refactor: 카테고리 페이지 구조 분리 및 개선 #12
